db setup
looked int how to setup sqlite and have a base form of the database
setup. the current process does the initialization on startup if it does
not find the database file. this will probably need to be moved later
but for now will suffice. the error handling process for the init setup
will need to be updated to better handle sql errors.