login update
this was fair bit of changes so here is a list
 - added in the argon2 crate for hashing passwords
 - moved the assets information into a separate config file
 - added in a default admin user when creating a new database
 - added in checks if there was an error when creating the database to
   properly handle cleaning up the database file.
 - the server will now properly close the database connection on
   shutdown
 - session creation will now actually generate a token and handle
   inserting duplicates (highly unlikely to happen)

the biggest change was adding in the login process for the server. on
request to the login page it will render a simple form and allow for
sending a username and password to the server. from there it will run
checks on the username and password then creating the session on the
happy path. on successful login the user will be redirected back to the
root page.