logout request
added in a logout request. for most of the cases when attempting to
initiator for the request it will just remove the session if it exists
or just remove the session id cookie. other instances will error out the
request. there is a temporary logout button in the dashboard but it will
be removed.