updated template loading
updated how the template files are retrieved. previousely they were
loaded one by one but this introduced an issue where a template file
could not be loaded because it extended another file that had not been
loaded yet. currently will use the add_template_files and give it a list
of the template files to load.