generate test data
this was to help with testing out the entries page.

added in logic for creating random test data for a user and added in
logic for creating extra users.

added in ids from another project to help with creating uids here. will
probably make changes from for this project and see how it goes.