journals
added in a journals table that will contain journal entries. not much in
terms of other logic right now. since I have not made a way to create
new journals the test users are created with a 'default' journal and the
current entries stuff uses that. later on I can add in logic to handle
multiple journals when requested and also handle a user sepcified
default journal if desired. can also use that as a point of reference
for access permissions when that comes up.