file entry updates
this is a large update. went through and added in the logic for sending
files to the server and cleaned up some of the processes for managing
the files. overall it works but encountered an issue with sqlite and
finding out that it does not support concurrent writes to the database.
will have to work on converting over to postgres for now.