postgres queries
this is incomplete but using this as a checkpoint of sorts. added in
functions that will use the postgres db connection without getting rid
of the sqlite ones for now to help with keeping track of errors and do a
more incremental update. the server builds but none of the handlers have
been updated to use postgres.