added in shadcn/ui
started work on adding in shadcn/ui code for styling and ui
interactivity. some of the documentation is not that great or non
existent so having to dig around in order to figure out what some of the
stuff requires so that it will work here. updated the login page to use
the new components and it looks better.