entry page update
updated the entry page to use the new styling. also added in a few extra
things to make the experience better when interacting with it. when
creating new video/audio recordings a dialog element will popup and you
can then iteract with the ui to start recording. once completed the
dialog will disappear and in order to view the new recording you can
select the view icon to bring up the dialog window from there. also
added in the ability to add files from the local system and to view
images.