updated test data generation
modified the test data generation to be outside the db setup. it is now
a command line argument and will happen after the state has been setup.
it will now properly create the journal directory so that file uploads
can happen without (hopefully) issue. also extended the possible range
for entry dates to go back to.