upload fix and form update
changed how the data is retrieved on the entry form page. it will now
use the form defaultValues async loading feature vs a speparate loading
state. it also remoes the need to have the useEffect to load the entry
from the server.

ran across an issue when uploading large amounts of files to the server.
not sure if it is the server or the browser having the issue but changed
the upload process to be file upload queue with async tasks that will
pull from the queue to upload the data.