journals update
added in the ability to interact with journals. you can now create
journals and being adding entries to it. all the necessary endpoints
have been updated to account for the journals id and no loger assume the
"default" journal. the ability to create journals is available but not
update and delete.