journal custom fields ui
added in the necessary interface components for adding custom fields to
a journal. also added in a new bundler to test out "rolldown". there are
a few issues with the build process that will prevent a successful build
but will be interested in seeing if it speeds up dev work.