journal custom fields modification
realized that I did not implement the fields in a way that I was
intending. previously you could change the field type without issue but
if there was any data that was created under that field then it would
now be incorrect. when a field is created it will be locked to that
specific config type.