sync entry #6
added in code to the sync endpoint to collect the user and journal and
create the entry in the database. the setup process is still manual and
will require modifying the database directly. some base tests have been
done and is able to create the new entries on the remote server.