finished file status #2
finished up the changes for the file status update. fixed issues found
along the way in the server side and cleaned up some of the logic so it
is a little easier to follow (hopefully). updated the front end to use
the new statuses and also fixed some previous issues with how the files
where handled. need to add in some more UI indicators for the uploads if
successful or failed and need to figure out how to handle the requested
state but with no file data available.