tail hash file upload #6
added in the ability to append a calculated hash of the uploaded data to
check against so that you do not have to calculate it before sending it
to the server. added in some simple tests to make sure the logic was
correct. still some cases that I have missed probably but is better than
nothing.