sync file data #6
added in the logic for syncing file data for entries. currently the
server will not transmit the actual file data to the peer. the peer will
mark the file entry as "Remote" to indicate the file exists but on the
peer server. I will add in the ability to send the actual file to the
server but for now this will work. need to update the file list so that
it will properly indicate a remote file. you will not be able to
download yet as more work will need to be added to allow retrieving file
data from the peer without downloading it to the local server.

also updated the file upload to only accept "Requested" files and any
file that is "Received" cannot be updated.