journal peer list #18
added in the ability to specify which peers the journal is to be sent
to. also added in the ability to tell the server to synchronize the
journal with all the peers. Note: the sync will not actually happen as
it will need to be reworked for the new method of sending data.