pki auth #19
this got out of hand (again). managed to get the authn working for the
api and client so now the peer server will know who is doing the
requests. there is a lot wrong with this and just going to commit what
I have and move on. will need to review all of this and do cleanup, fix,
and whatever else.