updated "sync" process #25
looked over what was being done previousely and updated the endpoints to
better match what I would like to do in the future. currently it is
still handled the same way as last time with adjustments to the file
entries portion. it is treated more in line with what the browser
endpoints do but does not returning any data. removed the previous
"sync" endpoint as it is no longer needed. did some testing and
everything seems to be working the same as previousely. still need to
upload the actual file to the peer at some point.