small changes to audio component #31
started to make some small changes the audio component to deal with the
errors from the ts type checking. there are a large number of them and
some may require reworking code to better handle the errors. those will
probably better to handle later on given how many their are.