updated entry route Errors #45
updated the remaining entry routes to use the new net::Error and create
the individual errors for each route. this required moving some code
around into different parts of the server but managed to get something
managable. `src/journals.rs` is getting big so at some point that should
be split up so it is not a total pain to use.