mfa recover front-end #15
added in the UI for enabling and disabling recovery codes for a user and
also to use a recovery code doring login when doing the MFA step.

started working on how to manage errors for the interface and updated
the auth page to use them. now some amount of errors will be displayed
to the user if someting goes wrong.