journals sidebar rework #51
modified the journals sidebar to be more of what I need it to be. will
allow for more easily accessible options in the future and looks better.

modified the top level sidebar to not make the profile picture request
if the value is undefined.

modified the root page view to display a message similar to the previous
journals page.