database migrations #53
added in a database migration that will be run through the tj2-cli.
modified the db directory to include a migrates subdirectory that will
house all the migrate files. init.sql was moved to V1__init.sql. will
have to come up with a way to easily rollback to a previous version when
working on new migrates if needed.