entries search and fmt
further updates to the entries search. while working on this I hit a few
snags with how to deal with the data in a table form. with some of the
searches I don't have a good way of returning snippets of the data while
also telling them how much is available. I have looked into some methods
of doing pagination so I will just have to try one.
 - removed the mod column fro mthe table
 - cargo fmt (includes other files)