entries calendar view
added in the calendar view for the entries. most of the core
fuctionality is there but some of the view stuff is not great. need to
figure out what how to display it when the view is small enough and what
to display if certain parts of the entry is not filled out.