file upload update
updated the file upload process to a new implementation. previously the
upload process was handled by a single function that was able to handle
the whole process but had a lot of logic to do multiple things at once.
the newer version use custom stream to manage the upload and also allows
for better individual testing of the components and a little easier to
read(?).
- removed the tests from `src/journal/custom_fields.rs` as they were all
  broken and will be reworked anyways.
- updated `blake3` to `1.8.2` to use newer features